+++ dask_geomodeling/geometry/base.py
@@ -17,7 +17,7 @@ class GeometryBlock(Block):
 
     A geometry request contains the following fields:
 
-    - mode: ``'intersects'`` or ``'extent'``. default ``'intersects'``.
+    - mode: one of ``{"intersects", "centroid", "extent"}``
     - geometry: limit returned objects to objects that intersect with this
       shapely geometry object
     - projection: projection to return the geometries in as WKT string
@@ -49,6 +49,41 @@ class GeometryBlock(Block):
     def set(self, *args):
         # NB cannot use __setitem__ as block instances are immutable
         return SetSeriesBlock(self, *args)
+
+    def to_file(self, *args, **kwargs):
+        """Utility function to export data from this block to a file on disk.
+
+        You need to specify the target file path as well as the extent geometry
+        you want to save.
+
+        Args:
+          url (str): The target file path. The extension determines the format.
+            For supported formats, consult
+            GeometryFileSink.supported_extensions.
+          fields (dict): a mapping that relates column names to output file
+            field names field names,
+            ``{<output file field name>: <column name>, ...}``.
+          tile_size (int): Optionally use this for large exports to stay within
+            memory constraints. The export is split in tiles of given size
+            (units are determined by the projection). Finally the tiles are
+            merged.
+          geometry (shapely Geometry): Limit exported objects to objects whose
+            centroid intersects with this geometry.
+          projection (str): The projection as a WKT string or EPSG code.
+            Sets the projection of the geometry argument, the target
+            projection of the data, and the tiling projection.
+          start (datetime): start date as UTC datetime
+          stop (datetime): stop date as UTC datetime
+          **request: see GeometryBlock request specification
+
+        Relevant settings can be adapted as follows:
+          >>> from dask import config
+          >>> config.set({"geomodeling.root": '/my/output/data/path'})
+          >>> config.set({"temporary_directory": '/my/alternative/tmp/dir'})
+        """
+        from dask_geomodeling.geometry.sinks import to_file
+
+        return to_file(self, *args, **kwargs)
 
 
 class SeriesBlock(Block):