+++ dask_geomodeling/utils.py
@@ -21,6 +21,12 @@ import fiona
 POLYGON = "POLYGON (({0} {1},{2} {1},{2} {3},{0} {3},{0} {1}))"
 
 
+try:
+    from fiona import Env as fiona_env  # NOQA
+except ImportError:
+    from fiona import drivers as fiona_env  # NOQA
+
+
 def get_index(values, no_data_value):
     """ Return an index to access for data values in values. """
     equal = np.isclose if values.dtype.kind == "f" else np.equal
@@ -331,7 +337,17 @@ def get_crs(user_input):
     """
     Return fiona CRS dictionary for user input.
     """
-    return fiona.crs.from_string(get_sr(user_input).ExportToProj4())
+    wkt = osr.GetUserInputAsWKT(str(user_input))
+    sr = osr.SpatialReference(wkt)
+    key = str("GEOGCS") if sr.IsGeographic() else str("PROJCS")
+    name = sr.GetAuthorityName(key)
+    if name == "EPSG":
+        # we can specify CRS in EPSG code which is more compatible with output
+        # file types
+        return fiona.crs.from_epsg(int(sr.GetAuthorityCode(key)))
+    else:
+        # we have to go through Proj4
+        return fiona.crs.from_string(sr.ExportToProj4())
 
 
 def crs_to_srs(crs):
@@ -369,6 +385,39 @@ def shapely_transform(geometry, src_srs, dst_srs):
     return shapely_wkb.loads(
         wkb_transform(geometry.wkb, src_sr=get_sr(src_srs), dst_sr=get_sr(dst_srs))
     )
+
+
+def geoseries_transform(df, src_srs, dst_srs):
+    """
+    Transform a GeoSeries to a different SRS. Returns a copy.
+
+    :param df: GeoSeries to transform
+    :param src_srs: source projection string
+    :param dst_srs: destination projection string
+
+    Note that we do not use .to_crs() for the transformation, because this is
+    much slower than OGR. Also, we ignore the .crs property (but we do set it)
+    """
+    result = df.geometry.apply(shapely_transform, args=(src_srs, dst_srs))
+    result.crs = get_crs(dst_srs)
+    return result
+
+
+def geodataframe_transform(df, src_srs, dst_srs):
+    """
+    Transform the geometry column of a GeoDataFrame to a different SRS.
+
+    :param df: GeoDataFrame to transform (will be changed inplace)
+    :param src_srs: source projection string
+    :param dst_srs: destination projection string
+
+    Note that we do not use .to_crs() for the transformation, because this is
+    much slower than OGR. Also, we ignore the .crs property (but we do set it)
+    """
+    geoseries = geoseries_transform(df.geometry, src_srs, dst_srs)
+    df.crs = geoseries.crs
+    df.set_geometry(geoseries)
+    return df
 
 
 def transform_min_size(min_size, geometry, src_srs, dst_srs):
@@ -684,7 +733,8 @@ def safe_file_url(url, start=None):
     if start is not None:
         warnings.warn(
             "Using the start argument in safe_file_url is deprecated. Use the "
-            "'geomodeling.root' in the dask config", DeprecationWarning
+            "'geomodeling.root' in the dask config",
+            DeprecationWarning,
         )
     else:
         start = config.get("geomodeling.root")